Move search function to API slice
